requirements.txt
SQLAlchemy==2.0.27
pymongo==4.6.1
pytest==8.0.2
matplotlib==3.8.3
plotly==5.19.0
python-dotenv==1.0.1
fastapi==0.109.2
uvicorn==0.27.1
pydantic==1.10.13
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
jinja2==3.1.3 